Finish Ejercicio-3 (Modulo_3)

+++ Logica_de_programacion/Ejercicios_Modulo_3/Ejercicio-3.cpp
@@ -31,7 +31,7 @@ int main()
     cout << "\n";
 
     // Creamos el array de alumnos
-    float *alumnos = new int[cantidad_alumnos];
+    float *alumnos = new float[cantidad_alumnos];
 
     // Cargamos las notas en el array
     for (int i = 1; i < cantidad_alumnos +1; i++)
